Прогнал линтер, исправил ошибки
