+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack --mode production",
-    "start": "webpack serve"
+    "start": "webpack serve",
+    "linter": "eslint ./ --fix"
   },
   "repository": {
     "type": "git",
