+++ src/pages/signup/lib/signupValidation.js
@@ -1,7 +1,7 @@
-import {validateEmail} from '../../../shared/validation/emailValidation.js';
-import {validateNickname} from '../../../shared/validation/nicknameValidation.js';
-import {validateBirthday} from '../../../shared/validation/birthdayValidation.js';
-import {validatePassword} from '../../../shared/validation/passwordValidation.js';
+import {validateEmail} from '../../../shared/validation/emailValidation';
+import {validateNickname} from '../../../shared/validation/nicknameValidation';
+import {validateBirthday} from '../../../shared/validation/birthdayValidation';
+import {validatePassword} from '../../../shared/validation/passwordValidation';
 
 
 /**
