--- src/shared/CSRF/CSRF.js
@@ -1,16 +1,0 @@
-class CSRF {
-    #csrfToken;
-    constructor() {
-        this.#csrfToken = '';
-    }
-
-    get() {
-        return this.#csrfToken;
-    }
-
-    set(csrf) {
-        this.#csrfToken = csrf;
-    }
-}
-
-export const csrf = new CSRF();
