+++ src/shared/components/authPage/authPageTemplate.hbs
