+++ src/shared/config/constants.js
@@ -1,7 +1,1 @@
-import { Auth } from '../../features/authorization/auth';
-
-export const constants = {
-    apiBaseUrl: '/api/v1/',
-};
-
-constants.auth = new Auth();
+export const API_BASE_URL = 'api/v1/';
